Рефакторинг

+++ src/main/java/by/gstu/fais/artyugin/education/model/dto/algorithmization/lab4/AlgLab4RequestDto.java
@@ -14,5 +14,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AlgLab4RequestDto {
-    private double[] numbers;
+    private int[] numbers;
 }
